RecentBlockController - query gas price on demand
